Rename prepare_data.prepare_data -> prepare_data.digits
