+++ lab/processes/pytorch_example/prepare_data/digits.py
